Caff parser
