+++ CaffParser/caff.h
@@ -1,45 +1,58 @@
 #pragma once
 #include "ciff.h"
 
-/*
-class Caff
-{
-private:
-    ull headerSize;
-    ull numAnim;
-    byte[] year;
-    byte month;
-    byte day;
-    byte hour;
-    byte minute;
-    ull creatorLen;
-    string creator;
-    ull duration;
-    vector<CIFF> ciffs;
-
-    Caff(ull hs, ull na, byte[]& ye, byte mo, byte da, byte ho, byte mi, ull crlen, const string& cr, ull dur, const vector<CIFF>& ci):
-        headerSize(hs), numAnim(na), year(ye), month(mo), day(da), hour(ho), minute(mi), creatorLen(crlen), creator(cr), duration(dur), ciffs(ci) {};
-
-    Caff(){};
-public:
-    static const int MIN_HEADER_SIZE = 4 + 8 + 8;
-    static const int MAIN_STAGES = 3;
-    static const int HEADER_STAGES = 3;
-    static const int CREDIT_STAGES = 7;
-    static const int ANIMATION_STAGES = 2;
-
-    inline ull getHeaderSize() { return headerSize; }
-    inline ull getNumAnim() { return numAnim; }
-    inline byte[] getYear() { return year; }
-    inline byte getMonth() { return month; }
-    inline byte getDay() { return day; }
-    inline byte getHour() { return hour; }
-    inline byte getMinute() { return minute; }
-    inline ull getCreatorLen() { return creatorLen; }
-    inline string getCreator() { return creator; }
-    inline ull getDuration() { return duration; }
-
-    CIFF getCIFF(ull);
-    static Caff parse(istream&, int , int);
+struct Credits {
+	int year;
+	int month;
+	int day;
+	int hour;
+	int minute;
+	string creator;
 };
-*/
+
+struct ParsingStatus {
+	bool headerDone;
+	bool creditsDone;
+	ull numAnims;
+	ull processedAnims;
+
+	ParsingStatus() : headerDone(false), creditsDone(false), numAnims(0), processedAnims(0) {}
+};
+
+struct Animation {
+	ull duration;
+	Ciff img;
+
+	Animation(): duration(0) {};
+};
+
+class Caff {
+private:
+    Credits credits;
+    vector<Animation> anims;
+
+    Caff(Credits c, const vector<Animation>& an):
+        credits(c), anims(an) {};
+
+    static void parseBlock(istream&, ull&, ParsingStatus&, Credits&, vector<Animation>&);
+    static ull parseHeader(istream&, const ull&);
+    static void parseCredits(istream&, const ull&, Credits&);
+    static void parseAnimation(istream&, const ull&, vector<Animation>&);
+
+public:
+	static const uint HEADER_SIZE = 20;
+	static const uint MIN_CREDITS_SIZE = 14;
+	static const uint MIN_ANIMATION_SIZE = Ciff::MIN_HEADER_SIZE + 8;
+    static const ull MAX_CAFF_SIZE = 209715200;
+
+    inline int getYear() const { return credits.year; }
+    inline int getMonth() const { return credits.month; }
+    inline int getDay() const { return credits.day; }
+    inline int getHour() const { return credits.hour; }
+    inline int getMinute() const { return credits.minute; }
+    inline string getCreator() const { return credits.creator; }
+    inline const vector<Animation>& getAnimations() const { return anims; }
+
+    const Ciff& getCiff(ull) const;
+    static Caff parse(istream&);
+};
