--- szg_bizt_proj/szg_bizt_proj.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30611.23
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "szg_bizt_proj", "szg_bizt_proj\szg_bizt_proj.vcxproj", "{58FBABFB-203C-427E-B294-0411FD16B27E}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{58FBABFB-203C-427E-B294-0411FD16B27E}.Debug|x64.ActiveCfg = Debug|x64
-		{58FBABFB-203C-427E-B294-0411FD16B27E}.Debug|x64.Build.0 = Debug|x64
-		{58FBABFB-203C-427E-B294-0411FD16B27E}.Debug|x86.ActiveCfg = Debug|Win32
-		{58FBABFB-203C-427E-B294-0411FD16B27E}.Debug|x86.Build.0 = Debug|Win32
-		{58FBABFB-203C-427E-B294-0411FD16B27E}.Release|x64.ActiveCfg = Release|x64
-		{58FBABFB-203C-427E-B294-0411FD16B27E}.Release|x64.Build.0 = Release|x64
-		{58FBABFB-203C-427E-B294-0411FD16B27E}.Release|x86.ActiveCfg = Release|Win32
-		{58FBABFB-203C-427E-B294-0411FD16B27E}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {1170F3E0-D178-4549-B2C0-9027EA5B0E61}
-	EndGlobalSection
-EndGlobal
