--- szg_bizt_proj/szg_bizt_proj/Debug/szg_bizt_proj.exe.recipe
@@ -1,7 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project>
-  <ProjectOutputs>C:\Users\Zsolt\source\repos\szg_bizt_proj\Debug\szg_bizt_proj.exe</ProjectOutputs>
-  <ContentFiles></ContentFiles>
-  <SatelliteDlls></SatelliteDlls>
-  <NonRecipeFileRefs></NonRecipeFileRefs>
-</Project>
