--- szg_bizt_proj/szg_bizt_proj/Debug/szg_bizt_proj.log
@@ -1,2 +1,0 @@
-﻿  szg_bizt_proj.cpp
-  szg_bizt_proj.vcxproj -> C:\Users\Zsolt\source\repos\szg_bizt_proj\Debug\szg_bizt_proj.exe
