--- szg_bizt_proj/szg_bizt_proj/Debug/szg_bizt_proj.tlog/szg_bizt_proj.lastbuildstate
@@ -1,2 +1,0 @@
-PlatformToolSet=v142:VCToolArchitecture=Native32Bit:VCToolsVersion=14.27.29110:TargetPlatformVersion=10.0.18362.0:
-Debug|Win32|C:\Users\Zsolt\source\repos\szg_bizt_proj\|
